requirements.txt

google_adk==1.12.0
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.7
kubernetes==28.1.0
google-cloud-logging==3.8.0
google-cloud-monitoring==2.16.0
structlog==23.2.0
python-dotenv==1.0.0
httpx==0.28.1
pytest==7.4.3
pytest-asyncio==0.21.1